OpenTaskDialog component has been made. Only thing missing is proper alignment of content, adding database stuff and adding the dialog to the Home view. The component requires two props the openDialog prop which opens the dialog and the task.
There are two emit but just one that will stay after database has been added. It is called closeDialog, and sends back a false, which closes the dialog.

That way the OpenTaskDialog can exist on Home view only one place, instead of adding it to every task there might be. The task can after all only be opened one at a time.

The TaskDialog was also added, and the certaintyDialog as well. These plus the steps and task checkmarks are the ones that are going to need to be hooked up to database.