The rest of the database requests were added, meaning everything functional works. It might not work perfect, but it works as desired. (#38)
Getters and setters were made for each form field, because they got a v-model which needed to not mutate the prop. Therefore getters and setters were added to mutate the prop, properly.

The openTaskDialog was by the database request added to the Home page and made work proberly. Some problems arose with the taskDialog which is why changes to the prop way of life, was made.